rename file to timeiNterval.ts
