+++ sdk/monitor/monitor-query/src/models/timeInterval.ts
